youtube/wikipedia/process/requirements.txt
beautifulsoup4==4.11.1
wikitextparser==0.49.2